# internal/features/users/transport/http/patch_user.go
package users_transport_http

import (
	"fmt"
	"net/http"
	"strings"

	"github.com/Alv1ol/Todoapp/internal/core/domain"
	core_logger "github.com/Alv1ol/Todoapp/internal/core/logger"
	core_http_request "github.com/Alv1ol/Todoapp/internal/core/transport/http/request"
	core_http_response "github.com/Alv1ol/Todoapp/internal/core/transport/http/response"
	core_http_types "github.com/Alv1ol/Todoapp/internal/core/transport/http/types"
	core_http_utils "github.com/Alv1ol/Todoapp/internal/core/transport/http/utils"
)

type PatchUserRequest struct {
	FullName    core_http_types.Nullable[string] `json:"full_name"`
	PhoneNumber core_http_types.Nullable[string] `json:"phone_number"`
}

func (r *PatchUserRequest) Validate() error{
	if r.FullName.Set{
		if r.FullName.Value == nil{
			return fmt.Errorf("`Fullname` can't be null")
		}

		fullNameLen := len([]rune(*r.FullName.Value))
		if fullNameLen < 3 || fullNameLen > 100 {
			return fmt.Errorf("`FullName` must be between 3 and 100 symbols")
		}
	}

	if r.FullName.Set {
		if r.PhoneNumber.Value != nil {
			phoneNumberLen := len([]rune(*r.PhoneNumber.Value))
			if phoneNumberLen < 10 || phoneNumberLen > 15 {
				return fmt.Errorf("`PhoneNumber` must be between 10 and 15 symbols")
			}

			if !strings.HasPrefix(*r.PhoneNumber.Value, "+"){
				return fmt.Errorf("`PhoneNumber` must starts with `+` symbol")
			}
		}
	}

	return nil
}

type PatchUserRespone UserDTOResponce

// PatchUser godoc
// @Summary Patch user
// @Description Patch user by id. All fields are optional, but at least one must be provided. FullName must be between 3 and 100 symbols. PhoneNumber must be between 10 and 15 symbols and starts with `+` symbol.
// @Tags users
// @Accept json
// @Produce json
// @Param id path int true "user id"
// @Param request body PatchUserRequest true "patch user request"
// @Success 200 {object} PatchUserRespone
// @Failure 400 {object} core_http_response.ErrorResponse "bad request"
// @Failure 404 {object} core_http_response.ErrorResponse "user not found"
// @Failure 500 {object} core_http_response.ErrorResponse "internal server error"
// @Router /users/{id} [patch]
func (h *UsersHTTPHandler) PathUser(rw http.ResponseWriter, r *http.Request) {
	ctx := r.Context()
	log := core_logger.FromContext(ctx)
	responseHandler := core_http_response.NewHTTPResponseHandler(log, rw)

	userID, err := core_http_utils.GetIntPathValue(r, "id")
	if err != nil {
		responseHandler.ErrorResponse(
			err,
			"failed to get userId path value",
		)
		return
	}

	var request PatchUserRequest
	if err := core_http_request.DecodeAndValidateRequest(r, &request); err != nil {
		responseHandler.ErrorResponse(
			err,
			"failed to decode and validate HTTP request",
		)

		return
	}

	userPatch := userPatchFromRequest(request)

	userDomain, err := h.usersService.PatchUser(ctx, userID, userPatch)
	if err != nil {
		responseHandler.ErrorResponse(
			err,
			"failed to patch user",
		)

		return
	}

	responce := PatchUserRespone(userDTOFromDomain(userDomain))

	responseHandler.JSONResponce(responce, http.StatusOK)
}

func userPatchFromRequest(request PatchUserRequest) domain.UserPatch {
	return domain.UserPatch{
		FullName:    request.FullName.ToDomain(),
		PhoneNumber: request.PhoneNumber.ToDomain(),
	}
}